Clean up refresh button
